+++ web/app/containers/cart/partials/cart-summary.jsx
@@ -10,13 +10,35 @@ import {CART_ESTIMATE_SHIPPING_MODAL} from '../constants'
 import {openModal} from 'progressive-web-sdk/dist/store/modals/actions'
 import {getSelectedShippingLabel, getPostcode} from '../../../store/checkout/shipping/selectors'
 import {getCheckoutShippingURL} from '../../app/selectors'
-import {removePromoCode} from '../actions' // @TODO figure out where this is coming from
+import {removePromoCode} from '../actions'
 
 import Button from 'progressive-web-sdk/dist/components/button'
 import CartPromoForm from './cart-promo-form'
 import Icon from 'progressive-web-sdk/dist/components/icon'
 import {Ledger, LedgerRow} from 'progressive-web-sdk/dist/components/ledger'
 import {Accordion, AccordionItem} from 'progressive-web-sdk/dist/components/accordion'
+
+// This is not written as a component to preserve the proptype
+// requirements of the Ledger component.
+const renderTaxAmountRow = (taxAmount, zipCode, openCalculateModal) => {
+    const editButton = (
+        <span>Based on delivery to
+            <Button innerClassName="u-padding-start-sm u-color-brand u-text-letter-spacing-normal" onClick={openCalculateModal}>
+                {zipCode}
+            </Button>
+        </span>
+    )
+
+    return (
+        <LedgerRow
+            className="u-flex-none u-border-0"
+            label="Taxes"
+            value={taxAmount}
+            labelAction={editButton}
+            key="Taxes"
+        />
+    )
+}
 
 const CartSummary = ({
     checkoutShippingURL,
@@ -38,27 +60,11 @@ const CartSummary = ({
         </Button>
     )
 
-    const editButton = (
-        <span>Based on delivery to
-            <Button innerClassName="u-padding-start-sm u-color-brand u-text-letter-spacing-normal" onClick={onCalculateClick}>
-                {zipCode}
-            </Button>
-        </span>
-    )
-
     const removeButton = (
         <Button innerClassName="u-color-brand u-padding-start-0 u-text-letter-spacing-normal" onClick={removePromoCode}>
             Remove Discount
         </Button>
     )
-
-    // when we want to show taxes with Calculate button
-    const neitherShippingNorDiscountsCalculatedYet = !zipCode && !discountLabel
-    const onlyDiscountCalculated = !zipCode && discountLabel
-
-    // When we want taxes to show with actual tax values
-    const onlyTaxIsCalculated = zipCode && !discountLabel
-    const bothTaxAndDiscountsCalculated = zipCode && discountLabel
 
     return (
         <div className="t-cart__summary">
@@ -97,18 +103,9 @@ const CartSummary = ({
                         />
                     }
 
-                    {(onlyTaxIsCalculated || bothTaxAndDiscountsCalculated) &&
-                        <LedgerRow
-                            className="u-flex-none u-border-0"
-                            label="Taxes"
-                            value={taxAmount}
-                            labelAction={editButton}
-                            key="Taxes"
-                        />
-                    }
-
-                    {(onlyDiscountCalculated || neitherShippingNorDiscountsCalculatedYet) &&
-                        <LedgerRow
+                    {taxAmount
+                        ? renderTaxAmountRow(taxAmount, zipCode, onCalculateClick)
+                        : <LedgerRow
                             className="u-flex-none"
                             label="Taxes"
                             labelAction="Rates based on shipping location"