+++ web/app/integration-manager/_merlins-connector/app/commands.js
@@ -9,7 +9,7 @@ import {getCart} from '../cart/commands'
 import {parseLoginStatus} from './parser'
 import {parseNavigation} from '../navigation/parser'
 import {receiveFormKey} from '../actions'
-import {CHECKOUT_SHIPPING_URL, CART_URL} from '../constants'
+import {CHECKOUT_SHIPPING_URL, CART_URL} from '../config'
 import {getCookieValue} from '../../../utils/utils'
 import {generateFormKeyCookie} from '../../../utils/magento-utils'
 
@@ -43,10 +43,7 @@ export const fetchPageData = (url) => (dispatch) => (
 
 export const initApp = () => (dispatch) => {
     // Use the pre-existing form_key if it already exists
-    let formKey = getCookieValue('form_key')
-    if (!formKey) {
-        formKey = generateFormKeyCookie()
-    }
+    const formKey = getCookieValue('form_key') || generateFormKeyCookie()
     dispatch(receiveFormKey(formKey))
 
     dispatch(setCheckoutShippingURL(CHECKOUT_SHIPPING_URL))