+++ web/app/integration-manager/_merlins-connector/checkout/parsers.js
@@ -2,14 +2,24 @@
 /* Copyright (c) 2017 Mobify Research & Development Inc. All rights reserved. */
 /* * *  *  * *  *  *  *  *  *  *  *  *  *  *  *  *  *  *  *  *  *  *  *  *  * */
 
+const removeUndefinedAddressKeys = (addressData) => {
+    for (const item in addressData) {
+        if (!addressData[item]) {
+            delete addressData[item]
+        }
+    }
+
+    return addressData
+}
+
 const getNameValue = (firstname, lastname) =>
       [firstname, lastname].filter((item) => item).join(' ')
 
 export const parseShippingInitialValues = (shippingFieldData) => {
     const fieldData = shippingFieldData.toJS()
     const streetFields = fieldData.street.children
-    return {
-        name: getNameValue(fieldData.firstname, fieldData.lastname),
+    const addressData = {
+        name: getNameValue(fieldData.firstname.value, fieldData.lastname.value),
         addressLine1: streetFields[0].value,
         company: fieldData.company.value,
         addressLine2: streetFields[1].value,
@@ -20,32 +30,50 @@ export const parseShippingInitialValues = (shippingFieldData) => {
         telephone: fieldData.telephone.value,
         billing_same_as_shipping: true
     }
+
+    // Remove undefined keys to prevent valid content being overriden in the store
+    return removeUndefinedAddressKeys(addressData)
 }
 
+/* eslint-disable camelcase */
 export const parseLocations = (shippingStepData) => {
     if (!shippingStepData) {
         return {}
     }
 
     return {
-        locations: {
-            countries: shippingStepData.getIn(['country_id', 'options']),
-            regions: shippingStepData.getIn(['region_id', 'options'])
-        }
+        countries: shippingStepData
+            .getIn(['country_id', 'options'])
+            .toJS()
+            .map(({value, label, is_region_required, is_zipcode_optional}) => ({
+                id: value,
+                label,
+                regionRequired: !!is_region_required,
+                postcodeRequired: !is_zipcode_optional
+            })),
+        regions: shippingStepData
+            .getIn(['region_id', 'options'])
+            // Skip the first option, since it's a placeholder.
+            .slice(1)
+            .toJS()
+            .map(({label, value, country_id}) => ({
+                id: value,
+                label,
+                countryId: country_id
+            }))
     }
 }
+/* eslint-enable camelcase */
 
 export const parseShippingMethods = (shippingMethods) => {
     if (!shippingMethods || !shippingMethods.map) {
         return []
     }
-    return shippingMethods.map((method) => {
-        return {
-            label: `${method.method_title} - ${method.carrier_title}`,
-            cost: `$${method.price_incl_tax.toFixed(2)}`,
-            value: `${method.carrier_code}_${method.method_code}`
-        }
-    })
+    return shippingMethods.map((method) => ({
+        label: `${method.method_title} - ${method.carrier_title}`,
+        cost: `$${method.price_incl_tax.toFixed(2)}`,
+        id: `${method.carrier_code}_${method.method_code}`
+    }))
 }
 
 export const checkoutConfirmationParser = ($, $html) => {