+++ web/app/integration-manager/app/commands.js
@@ -11,5 +11,6 @@ export const register = (commands) => {
 /**
  * Initializes the connector during app startup. This command dispatched
  * be called before any other integration manager commands are.
+ * @function
  */
 export const initApp = () => connector.initApp()