+++ web/app/integration-manager/cart/commands.js
@@ -9,6 +9,7 @@ export const register = (commands) => {
 }
 /**
  * Initializes any required data for the Cart page
+ * @function
  * @param {string} url The url of the current page
  * @param {string} routeName The route name of the current page
  */
@@ -16,11 +17,13 @@ export const initCartPage = (url, routeName) => connector.initCartPage(url, rout
 
 /**
  * Retrieves the current cart information.
+ * @function
  */
 export const getCart = () => connector.getCart()
 
 /**
  * Adds a product to the cart
+ * @function
  * @param productId {string} The product's ID
  * @param quantity {number} The quantity to add
  */
@@ -28,12 +31,14 @@ export const addToCart = (productId, quantity) => connector.addToCart(productId,
 
 /**
  * Removes an item from the cart
+ * @function
  * @param itemID {string} The cart item ID to remove
  */
 export const removeFromCart = (itemID) => connector.removeFromCart(itemID)
 
 /**
  * Updates the quantity of the given item in the cart
+ * @function
  * @param itemID {string} The cart item ID to update
  * @param quantity {number} The new quantity
  */
@@ -41,6 +46,7 @@ export const updateItemQuantity = (itemID, quantity) => connector.updateItemQuan
 
 /**
  * Add a product to the wishlist
+ * @function
  * @param productId {string} The product's ID
  * @param productURL {strin} The URL of the product's detail page
  */
@@ -48,6 +54,7 @@ export const addToWishlist = (productId, productURL) => connector.addToWishlist(
 
 /**
  * Estimates taxes for a proposed address and shipping method
+ * @function
  * @param address {object} The address to use for tax estimation
  * @param shippingMethod {string} The shipping method to use for tax estimation (connector-specific!)
  */
@@ -55,12 +62,14 @@ export const fetchTaxEstimate = (address, shippingMethod) => connector.fetchTaxE
 
 /**
  * Adds promo code
+ * @function
  * @param couponCode {string} The coupon code to be applied
  */
 export const putPromoCode = (couponCode) => connector.putPromoCode(couponCode)
 
 /**
  * Deletes promo code
+ * @function
  * @param couponCode {string} The coupon code to be removed
  */
 export const deletePromoCode = (couponCode) => connector.deletePromoCode(couponCode)