+++ web/app/integration-manager/commands.js
@@ -9,6 +9,8 @@ import {register as registerCart} from './cart/commands'
 import {register as registerApp} from './app/commands'
 import {register as registerCheckout} from './checkout/commands'
 import {register as registerAccount} from './account/commands'
+import {register as registerCustomCommands} from './custom/commands'
+
 
 let connector = {}
 
@@ -23,5 +25,12 @@ export const register = (commands) => {
     registerAccount(commands.account)
 }
 
+export const registerCustom = (commands) => {
+    registerCustomCommands(commands)
+}
+
+/** @function */
 export const submitNewsletter = (...args) => connector.submitNewsletter(...args)
+
+/** @function */
 export const getSearchSuggestions = (...args) => connector.getSearchSuggestions(...args)