+++ web/app/store/checkout/reducer.js
@@ -4,18 +4,27 @@
 
 import Immutable from 'immutable'
 import {handleActions} from 'redux-actions'
-import {mergePayload} from '../../utils/reducer-utils'
+import {mergePayload, setCustomContent} from '../../utils/reducer-utils'
 import {receiveSavedShippingAddresses} from './actions'
 import * as integrationManagerResults from '../../integration-manager/checkout/results'
 import {setDefaultShippingAddressId} from './shipping/actions'
 
 const checkoutReducer = handleActions({
     [receiveSavedShippingAddresses]: mergePayload,
-    [integrationManagerResults.receiveBillingInitialValues]: mergePayload,
-    [integrationManagerResults.receiveShippingInitialValues]: mergePayload,
+    [integrationManagerResults.receiveCheckoutLocations]: mergePayload,
+    [integrationManagerResults.receiveBillingAddress]: mergePayload,
+    [integrationManagerResults.receiveShippingAddress]: mergePayload,
     [integrationManagerResults.receiveCheckoutData]: mergePayload,
     [integrationManagerResults.receiveUserEmail]: mergePayload,
     [integrationManagerResults.receiveCheckoutCustomContent]: mergePayload,
+    [integrationManagerResults.receiveLocationsCustomContent]: setCustomContent('locations'),
+    [integrationManagerResults.receiveShippingAddressCustomContent]: setCustomContent('shippingAddress'),
+    [integrationManagerResults.receiveBillingAddressCustomContent]: setCustomContent('billingAddress'),
+    [integrationManagerResults.receiveShippingMethods]: (state, {payload}) => (
+        // Using `set` here will make sure the list in the store is
+        // correctly truncated.
+        state.set('shippingMethods', Immutable.fromJS(payload))
+    ),
     [setDefaultShippingAddressId]: mergePayload
 }, Immutable.Map())