+++ web/app/store/checkout/shipping/selectors.js
@@ -6,14 +6,15 @@ import {createSelector} from 'reselect'
 import Immutable from 'immutable'
 import {createGetSelector} from 'reselect-immutable-helpers'
 import {getCheckout} from '../../selectors'
+import {getShippingMethods} from '../selectors'
 
 export const getShipping = createGetSelector(getCheckout, 'shipping', Immutable.Map())
 
+export const getShippingCustomContent = createGetSelector(getShipping, 'custom')
+
 export const getSavedAddresses = createGetSelector(getCheckout, 'savedAddresses', Immutable.List())
 
-export const getShippingMethods = createGetSelector(getShipping, 'shippingMethods', Immutable.List())
-
-export const getShippingAddress = createGetSelector(getShipping, 'address', Immutable.Map())
+export const getShippingAddress = createGetSelector(getCheckout, 'shippingAddress', Immutable.Map())
 
 export const getInitialShippingAddress = createSelector(
     getCheckout,
@@ -36,7 +37,7 @@ export const getSelectedShippingMethod = createSelector(
         if (!shippingMethods.size) {
             return Immutable.Map()
         }
-        const selectedValue = shippingMethods.filter((method) => method.get('value') === selectedMethodValue)
+        const selectedValue = shippingMethods.filter((method) => method.get('id') === selectedMethodValue)
         return selectedValue.size ? selectedValue.get(0) : shippingMethods.get(0)
     })
 
@@ -67,3 +68,6 @@ export const getRegionId = createGetSelector(getShippingAddress, 'regionId')
 export const getCountryId = createGetSelector(getShippingAddress, 'countryId')
 
 export const getCity = createGetSelector(getShippingAddress, 'city')
+
+export const getShippingAddressCustomContent = createGetSelector(getShippingAddress, 'custom')
+