+++ web/app/utils/utils.js
@@ -47,7 +47,7 @@ export const splitFullName = (fullname) => {
 
 /**
  * Currently requestIdleCallback is only supported in Chrome,
- * we'll have to provide a fallback for iOS Safari
+ * TODO: We'll have to provide a fallback for iOS Safari
  * https://developers.google.com/web/updates/2015/08/using-requestidlecallback
  * http://caniuse.com/#feat=requestidlecallback
  */
@@ -68,43 +68,25 @@ export const typecheck = (type, value) => {
     return value
 }
 
-export const createTypedAction = (description, type) => createReduxAction(
+/**
+ * Create an action creator that typechecks its argument.
+ *
+ * The action creator argument is passed unchanged as the payload if
+ * no key is passed, while if a key is provided the action creator
+ * argument is wrapped in an object under that key. This allows the
+ * action to set a specific key within the Redux store using mergePayload.
+ *
+ * @param description {string} The description of the action (seen in dev tools)
+ * @param type {Runtype} The type to check the action argument against
+ * @param key {string} (optional) The key in the store to set with the payload
+ * @returns {function} The action creator.
+ */
+export const createTypedAction = (description, type, key) => createReduxAction(
     description,
-    (payload) => typecheck(type, payload)
+    key
+        ? (payload) => { return {[key]: typecheck(type, payload)} }
+        : (payload) => typecheck(type, payload)
 )
-
-export const parseLocationData = (formValues, registeredFieldNames) => {
-    // Default values to use if none have been selected
-    const address = {country_id: 'US', region_id: '0', postcode: null}
-
-    if (formValues) {
-        // Only return the field value if the field is registered
-        const getRegisteredFieldValue = (fieldName) => {
-            return registeredFieldNames.includes(fieldName) ? formValues[fieldName] : undefined
-        }
-
-        const countryId = getRegisteredFieldValue('country_id')
-        if (countryId) {
-            address.country_id = countryId
-        }
-
-        const postcode = getRegisteredFieldValue('postcode')
-        if (postcode) {
-            address.postcode = postcode
-        }
-
-        if (formValues.region) {
-            address.region = getRegisteredFieldValue('region')
-            // Remove the region_id in case we have an old value
-            delete address.region_id
-        } else {
-            address.region_id = getRegisteredFieldValue('region_id')
-        }
-    }
-
-    return address
-}
-
 
 export const buildQueryString = (query) => {
     return query.replace(/ /g, '+')