+++ web/tests/system/page-objects/cart.js
@@ -3,11 +3,14 @@
 /* * *  *  * *  *  *  *  *  *  *  *  *  *  *  *  *  *  *  *  *  *  *  *  *  * */
 
 const selectors = {
+    miniCart: '.qa-header__cart',
+    viewCart: '.t-mini-cart__content .c--tertiary',
     cartTemplateIdentifier: '.t-cart.t--loaded',
     cartCheckout: '.qa-cart__checkout',
     removeItem: '.qa-cart__remove-item',
     confirmRemove: '.t-cart__remove-item-confirmation-modal .c--secondary',
-    emptyCart: '.t-cart__empty'
+    emptyCart: '.t-cart__empty',
+    emptyMiniCart: '.t-mini-cart__empty-content'
 }
 
 const Cart = function(browser) {
@@ -29,21 +32,27 @@ Cart.prototype.removeItems = function() {
     // Remove all items from the cart
     const self = this
     this.browser
-        .log('Removing item')
-        .url('https://www.merlinspotions.com/checkout/cart/')
-        .waitForElementVisible(selectors.cartTemplateIdentifier)
-        .element('css selector', selectors.removeItem, (result) => {
+        .preview()
+        .log('Opening mini cart')
+        .waitForElementVisible(selectors.miniCart)
+        .click(selectors.miniCart)
+        .waitForAnimation()
+        .element('css selector', selectors.viewCart, (result) => {
             if (result.value && result.value.ELEMENT) {
                 self.browser
+                    .log('View cart')
+                    .waitForElementVisible(selectors.viewCart)
+                    .click(selectors.viewCart)
                     .log('Removing item from cart')
-                    .click(selectors.removeItem)
+                    .execute(`document.querySelector('${selectors.removeItem}').click()`)
                     .waitForElementVisible(selectors.confirmRemove)
-                    .click(selectors.confirmRemove)
+                    .execute(`document.querySelector('${selectors.confirmRemove}').click()`)
                     .waitUntilMobified()
+                    .waitForElementVisible(selectors.emptyCart)
                 self.removeItems()
             }
         })
-        .waitForElementVisible(selectors.emptyCart)
+        .waitForElementVisible(selectors.emptyMiniCart)
     return this
 }