+++ web/webpack/base.loader.js
@@ -60,8 +60,13 @@ module.exports = {
             }
         }),
         new webpack.DefinePlugin({
+            // This is defined as a boolean, not a string
+            MESSAGING_ENABLED: `${webPackageJson.messagingEnabled}`,
+            // These are defined as string constants
             MESSAGING_SITE_ID: `'${webPackageJson.messagingSiteId}'`,
-            NATIVE_WEBPACK_ASTRO_VERSION: readNativeAstroVersion()
+            NATIVE_WEBPACK_ASTRO_VERSION: readNativeAstroVersion(),
+            PROJECT_SLUG: `'${webPackageJson.projectSlug}'`,
+            AJS_SLUG: `'${webPackageJson.aJSSlug}'`
         })
     ]
 }